+++ library/src/schemas/record/types.ts
@@ -5,6 +5,11 @@ import type {
   Output,
   ResolveObject,
 } from '../../types.ts';
+import type { EnumSchema, EnumSchemaAsync } from '../enumType/index.ts';
+import type {
+  NativeEnumSchema,
+  NativeEnumSchemaAsync,
+} from '../nativeEnum/index.ts';
 import type { RecordKey } from './record.ts';
 import type { RecordKeyAsync } from './recordAsync.ts';
 
@@ -19,12 +24,25 @@ export type RecordPathItem = {
 };
 
 /**
+ * Any enum schema type.
+ */
+type AnyEnumSchema =
+  | EnumSchema<any>
+  | EnumSchemaAsync<any>
+  | NativeEnumSchema<any>
+  | NativeEnumSchemaAsync<any>;
+
+/**
  * Record input inference type.
  */
 export type RecordInput<
   TRecordKey extends RecordKey | RecordKeyAsync,
   TRecordValue extends BaseSchema | BaseSchemaAsync
-> = ResolveObject<Record<Input<TRecordKey>, Input<TRecordValue>>>;
+> = ResolveObject<
+  TRecordKey extends AnyEnumSchema
+    ? Partial<Record<Input<TRecordKey>, Input<TRecordValue>>>
+    : Record<Input<TRecordKey>, Input<TRecordValue>>
+>;
 
 /**
  * Record output inference type.
@@ -32,4 +50,8 @@ export type RecordInput<
 export type RecordOutput<
   TRecordKey extends RecordKey | RecordKeyAsync,
   TRecordValue extends BaseSchema | BaseSchemaAsync
-> = ResolveObject<Record<Output<TRecordKey>, Output<TRecordValue>>>;
+> = ResolveObject<
+  TRecordKey extends AnyEnumSchema
+    ? Partial<Record<Output<TRecordKey>, Output<TRecordValue>>>
+    : Record<Output<TRecordKey>, Output<TRecordValue>>
+>;